pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-ai"
version = "0.1.0"
description = "Production-grade agentic AI system"
requires-python = ">=3.11"
dependencies = [
  "langchain>=1.2,<2",
  "langchain-ollama>=1,<2",
  "langchain-core>=1.2,<2",
  "langchain-community>=0.4,<1",

  "fastapi>=0.128,<1",
  "uvicorn[standard]>=0.40,<1",
  "pydantic>=2,<3",

  "python-dotenv>=1,<2",
  "httpx>=0.28,<1",
  "tenacity>=9,<10",

  "streamlit>=1.53,<2",
  "apscheduler>=3.11,<4",

  "sqlalchemy>=2,<3",
  "sqlite-utils>=3.39,<4",

  "llama-index>=0.14,<1"
]
